Super refactoring going down, bro
