+++ src/civilization/census/components.rs
@@ -1,0 +1,9 @@
+use bevy::prelude::{Component, Reflect};
+
+#[derive(Component, Debug, Reflect)]
+pub struct Census {
+    pub population: usize,
+}
+
+#[derive(Component, Debug, Reflect)]
+pub struct HasPopulation;
