+++ src/civilization/census/events.rs
@@ -1,0 +1,7 @@
+use bevy::prelude::Event;
+
+#[derive(Event, Debug)]
+pub struct CensusEnded;
+
+#[derive(Event, Debug)]
+pub struct CheckAreasForPopulationCommand;
