+++ src/civilization/census/mod.rs
@@ -1,0 +1,5 @@
+pub mod plugin;
+pub mod systems;
+pub mod events;
+pub mod components;
+pub mod resources;
