+++ src/civilization/mod.rs
@@ -1,5 +1,5 @@
-pub mod civ;
 pub mod population_expansion;
-pub mod census;
 pub mod movement;
 pub mod game_phases;
+pub mod census;
+pub mod general;
