+++ src/lib.rs
@@ -18,7 +18,7 @@ use bevy::app::App;
 // #[cfg(debug_assertions)]
 // use bevy::diagnostic::{FrameTimeDiagnosticsPlugin, LogDiagnosticsPlugin};
 use bevy::prelude::*;
-use crate::civilization::civ::CivilizationPlugin;
+use crate::civilization::general::plugin::CivilizationPlugin;
 
 // This example game uses States to separate logic
 // See https://bevy-cheatbook.github.io/programming/states.html
